fixed nvidia installer
added check if nvidia card is used in container

+++ start.sh
@@ -1,48 +1,45 @@
 #!/bin/bash
+echo "---Checking for optional scripts---"
+if [ -f /opt/scripts/user.sh ]; then
+	echo "---Found optional script, executing---"
+    chmod +x /opt/scripts/user.sh
+    /opt/scripts/user.sh
+else
+	echo "---No optional script found, continuing---"
+fi
+
 export DATA_DIR=$HOME
 
-if [ -z "${NV_DRV_V}" ]; then
-        echo "---Trying to get Nvidia driver version---"
-        export NV_DRV_V="$(nvidia-smi | grep NVIDIA-SMI | cut -d ' ' -f3)"
+if [ -f /usr/bin/nvidia-smi ]; then
+        echo "---Detected Nvidia card, installing driver, please wait!---"
         if [ -z "${NV_DRV_V}" ]; then
-                echo "---Something went wrong, can't get driver version, putting container into sleep mode---"
-                sleep infinity
-        else
-                echo "---Successfully got driver version: ${NV_DRV_V}---"
+                echo "---Trying to get Nvidia driver version---"
+                export NV_DRV_V="$(nvidia-smi | grep NVIDIA-SMI | cut -d ' ' -f3)"
+                if [ -z "${NV_DRV_V}" ]; then
+                        echo "---Something went wrong, can't get driver version, putting container into sleep mode---"
+                        sleep infinity
+                else
+                        echo "---Successfully got driver version: ${NV_DRV_V}---"
+                fi
         fi
-fi
-
-if [ ! -z "$INSTALL_V" ]; then
-        if [ "$INSTALL_V" != "${NV_DRV_V}" ]; then
-                echo "---Version missmatch, deleting local Nvidia Driver v$INSTALL_V---"
-                rm ${DATA_DIR}/NVIDIA_$INSTALL_V.run
+        
+        if [ ! -z "$INSTALL_V" ]; then
+                if [ "$INSTALL_V" != "${NV_DRV_V}" ]; then
+                        echo "---Version missmatch, deleting local Nvidia Driver v$INSTALL_V---"
+                        rm ${DATA_DIR}/NVIDIA_$INSTALL_V.run
+                fi
         fi
-fi
 
-INSTALL_V="$(find ${DATA_DIR} -name NVIDIA_*\.run | cut -d '_' -f 2 | cut -d '.' -f1- | sed 's/\.[^.]*$//')"
+        INSTALL_V="$(find ${DATA_DIR} -name NVIDIA_*\.run | cut -d '_' -f 2 | cut -d '.' -f1- | sed 's/\.[^.]*$//')"
 
-if [ ! -z "$INSTALL_V" ]; then
-        if [ "$INSTALL_V" != "${NV_DRV_V}" ]; then
-                echo "---Version missmatch, deleting local Nvidia Driver v$INSTALL_V---"
-                rm ${DATA_DIR}/NVIDIA_$INSTALL_V.run
+        if [ ! -z "$INSTALL_V" ]; then
+                if [ "$INSTALL_V" != "${NV_DRV_V}" ]; then
+                        echo "---Version missmatch, deleting local Nvidia Driver v$INSTALL_V---"
+                        rm ${DATA_DIR}/NVIDIA_$INSTALL_V.run
+                fi
         fi
-fi
 
-if [ ! -f /usr/bin/nvidia-settings ]; then
-        if [ -f ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ]; then
-                echo "---Found NVIDIA Driver v${NV_DRV_V} locally, installing...---"
-                ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ${NVIDIA_BUILD_OPTS} >/dev/null 2>&1
-        else
-                echo "---Downloading and installing Nvidia Driver v${NV_DRV_V}---"
-                wget -q --show-progress --progress=bar:force:noscroll -O /tmp/NVIDIA.run http://download.nvidia.com/XFree86/Linux-x86_64/${NV_DRV_V}/NVIDIA-Linux-x86_64-${NV_DRV_V}.run &&
-                        chmod +x /tmp/NVIDIA.run &&
-                        /tmp/NVIDIA.run ${NVIDIA_BUILD_OPTS} >/dev/null 2>&1 &&
-                        mv /tmp/NVIDIA.run ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run
-        fi
-else
-        CUR_NV_DRV_V=$INSTALL_V
-        if [ "$NV_DRV_V" != "$CUR_NV_DRV_V" ]; then
-                echo "---Driver version missmatch, currently installed: v$CUR_NV_DRV_V, driver on Host: v$NV_DRV_V---"
+        if [ ! -f /usr/bin/nvidia-settings ]; then
                 if [ -f ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ]; then
                         echo "---Found NVIDIA Driver v${NV_DRV_V} locally, installing...---"
                         ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ${NVIDIA_BUILD_OPTS} >/dev/null 2>&1
@@ -54,9 +51,28 @@ else
                                 mv /tmp/NVIDIA.run ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run
                 fi
         else
-                echo "---Nvidia Driver v$CUR_NV_DRV_V Up-To-Date---"
+                CUR_NV_DRV_V=$INSTALL_V
+                if [ "$NV_DRV_V" != "$CUR_NV_DRV_V" ]; then
+                        echo "---Driver version missmatch, currently installed: v$CUR_NV_DRV_V, driver on Host: v$NV_DRV_V---"
+                        if [ -f ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ]; then
+                                echo "---Found NVIDIA Driver v${NV_DRV_V} locally, installing...---"
+                                ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run ${NVIDIA_BUILD_OPTS} >/dev/null 2>&1
+                        else
+                                echo "---Downloading and installing Nvidia Driver v${NV_DRV_V}---"
+                                wget -q --show-progress --progress=bar:force:noscroll -O /tmp/NVIDIA.run http://download.nvidia.com/XFree86/Linux-x86_64/${NV_DRV_V}/NVIDIA-Linux-x86_64-${NV_DRV_V}.run &&
+                                chmod +x /tmp/NVIDIA.run &&
+                                /tmp/NVIDIA.run ${NVIDIA_BUILD_OPTS} >/dev/null 2>&1 &&
+                                mv /tmp/NVIDIA.run ${DATA_DIR}/NVIDIA_${NV_DRV_V}.run
+                        fi
+                else
+                        echo "---Nvidia Driver v$CUR_NV_DRV_V Up-To-Date---"
+                fi
         fi
 fi
+
+echo "---Setting permissions---"
+chown -R root:100 /opt/scripts
+chmod -R 750 /opt/scripts
 
 term_handler() {
         kill -SIGTERM "$killpid"
